chore(platform): thin out platform client
Most of the platform related functions are outside the client already.
Let's thin out the client functionality and by that make the interface
smaller. Furthermore, remove methods of the interface that are not
called anywhere.
