+++ pkg/cli/list_helm.go
@@ -83,7 +83,7 @@ func printVClusters(ctx context.Context, options *ListOptions, output []ListVClu
 				ctx, cancel := context.WithTimeout(ctx, 2*time.Second)
 				defer cancel()
 
-				proVClusters, _ := platformClient.ListVClusters(ctx, "", "")
+				proVClusters, _ := platform.ListVClusters(ctx, platformClient, "", "")
 				if len(proVClusters) > 0 {
 					logger.Infof("You also have %d virtual clusters in your platform manager context.", len(proVClusters))
 					logger.Info("If you want to see them, run: 'vcluster list --manager platform' or 'vcluster use manager platform' to change the default")
